+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## - 2026-09-01
+### Añadido
+- Tests de integración para proyecto aumentados.
+- Corregido un bug por el que al obtener "mis proyectos" daba not found si no tenias proyectos (ahora muestra un Ok con una lista vacia).
+- Corregido un bug por el que al obtener los proyectos de usuario daba not found si no tenia proyectos (ahora muestra un Ok con una lista vacia).
+
+
 ## - 2026-08-28
 ### Añadido
 - Inicio para Test de Integración.